+++ agijagi/app/src/main/java/likelion/project/agijagi/chatting/ChattingRoomFragment.kt
@@ -1,4 +1,4 @@
-package likelion.project.agijagi.userchat
+package likelion.project.agijagi.chatting
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,14 +7,14 @@ import android.view.View
 import android.view.ViewGroup
 import likelion.project.agijagi.R
 
-class UserChatRoomFragment : Fragment() {
+class ChattingRoomFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_user_chat_room, container, false)
+        return inflater.inflate(R.layout.fragment_chatting_room, container, false)
     }
 
 }
