+++ agijagi/app/src/main/res/layout/fragment_chatting_room.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".userchat.UserChatRoomFragment">
+    tools:context=".chatting.ChattingRoomFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
